Refactored naming of section to dataframe_sql
